# dashboard/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Studio OS API",
  "description": "Shared coordination API for Studio OS projects: tasks, resource claims, work sessions, decisions, AI work logs, events, and file transfer metadata. Every operation under `/api/v1` requires machine authentication (`Authorization: Bearer <machine-token>`, provisioned out of band) except `GET /healthz`, `GET /metrics` and the human dashboard login `POST /auth/token`. File bytes never flow through this API \u2014 transfers exchange metadata and short-lived signed URLs only, uploads and downloads go directly to object storage. Replayable creations accept `Idempotency-Key`; concurrent updates use `If-Match-Version`; permission refusals answer `403 forbidden` and are final.",
  "version": "1"
 },
 "paths": {
...
  "/api/v1/auth/token": {
   "post": {
    "tags": [
     "auth"
    ],
    "summary": "Login",
    "description": "Exchange human credentials for a short-lived dashboard JWT.",
    "operationId": "login_api_v1_auth_token_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/TokenRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TokenResponse"
        }
       }
      }
     },
     "401": {
      "description": "Invalid email or password."
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "TokenRequest": {
    "properties": {
     "email": {
      "type": "string",
      "title": "Email",
      "examples": [
       "admin@example.com"
      ]
     },
     "password": {
      "type": "string",
      "title": "Password",
      "examples": [
       "super-secret"
      ]
     }
    },
    "type": "object",
    "required": [
     "email",
     "password"
    ],
    "title": "TokenRequest"
   },
   "TokenResponse": {
    "properties": {
     "access_token": {
      "type": "string",
      "title": "Access Token"
     },
     "token_type": {
      "type": "string",
      "title": "Token Type",
      "default": "bearer"
     }
    },
    "type": "object",
    "required": [
     "access_token"
    ],
    "title": "TokenResponse"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  },
  "securitySchemes": {
   "MachineBearer": {
    "type": "http",
    "description": "Machine credential provisioned out of band. Send it as `Authorization: Bearer <machine-token>` on every request under `/api/v1`. The token is opaque: a missing, invalid or revoked credential returns 401. Unauthenticated operations are `GET /healthz`, `GET /metrics` and the human dashboard login `POST /auth/token`.",
    "scheme": "bearer"
   }
  }
 },
...
}